+++ art/md/cn/CHANGELOG.md
@@ -90,7 +90,7 @@
 ## V 1.1.9
 >修复：不设置loadMore或者onRefresh时，回拉异常的问题  
 
-## V 下个版本
+## V 1.2.0
 >修复：onNoMore回调不准确问题  
 >添加：isRefreshing用于判断是否正在刷新或加载中  
 >添加：Header和Footer状态改变回调